% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.clr.r
\name{omicplotr.clr}
\alias{omicplotr.clr}
\title{Generate prcomp object to plot compositional biplot with
 \code{omicplotr.colouredPCA}}
\usage{
omicplotr.clr(x, var.filt = 0)
}
\arguments{
\item{x}{filtered data frame by \code{omicplotr.countfilter}}

\item{var.filt}{variance cutoff to be applied after clr is computed,
default, 0.}
}
\description{
\code{omciplotr.clr} replaces zeros using CZM method from
 \code{zCompositions}, calculates CLR, filters by variance and
 generates a \code{prcomp} object from the transformed data frame.
}
