% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.colouredPCA.r
\name{omicplotr.colouredPCA}
\alias{omicplotr.colouredPCA}
\title{Generate coloured compositional biplot.}
\usage{
omicplotr.colouredPCA(x, colourvector, scale = 0, arrows = TRUE,
  taxonomy = NULL, show.taxonomy = FALSE, tax.level = 6,
  removenames = FALSE, names.cex = 1, points.cex = 0.8,
  main = "Principal\\n     Component Analysis Biplot")
}
\arguments{
\item{x}{prcomp object generated by \code{omicplotr.clr}.}

\item{colourvector}{list of colours generated by \code{omciplotr.colvec}.}

\item{scale}{0 = relationship between samples, 1 = relationship between rows
default 0.}

\item{arrows}{show PCA arrows, default TRUE.}

\item{taxonomy}{taxonomy column from \code{omicplotr.filter} to plot points
as taxonomic identifiers.}

\item{show.taxonomy}{show taxonomy as points, default 0. Required input is
x$taxonomy.}

\item{tax.level}{numeric taxonomic level (kingdom =1, phylum =2...) to show
if \code{show.taxonomy = TRUE}, default 6 (genus).}

\item{removenames}{replace sample names with points, default FALSE.}

\item{names.cex}{the size of sample names.}

\item{points.cex}{the size of points representing features (text size for
taxonomic levels if \code{show.taxonomy = TRUE}).}
}
\description{
Generates a coloured PCA biplot from a \code{prcomp} object
 generated by \code{omicplotr.clr} using \code{coloredBiplot}
 from the \code{compositions} package.
}
