% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.filter.r
\name{omicplotr.filter}
\alias{omicplotr.filter}
\title{Filter data frame by read counts.}
\usage{
omicplotr.filter(x, min.reads = 0, min.count = 0, min.sum = 0,
  min.prop = 0, max.prop = 1)
}
\arguments{
\item{x}{counts table as data frame.}

\item{min.reads}{removes any columns with sum less than or equal to input.}

\item{min.count}{removes any rows with maximum lower than input.}

\item{min.sum}{removes any rows with sum lower than input.}

\item{min.prop}{minimum proportional abundance in any sample.}

\item{max.prop}{maximum proportional abundance in any sample.}
}
\description{
\code{omicplotr.filter} removes columns and/or rows of table below threshold.
This must be used after \code{omicplotr.metadataFilter}.
}
\examples{
# The 'reads' data frame must have unique row and column names, and may
# include a taxonomy column as the last column, and look like this
# (from \\code{ALDEx2} documentation):
#
#              T1a T1b  T2  T3  N1  N2  Nx taxonomy
#   Gene_00001   0   0   2   0   0   1   0 Bacteria;Bacteroidetes;...
#   Gene_00002  20   8  12   5  19  26  14 Bacteria;Firmicutes;...
#   Gene_00003   3   0   2   0   0   0   1 Bacteria;Proteobacteria;...
#   Gene_00004  75  84 241 149 271 257 188 Bacteria;Bacteroidetes;...
#   Gene_00005  10  16   4   0   4  10  10 Bacteria;unclassified;...
#   Gene_00006 129 126 451 223 243 149 209 Bacteria;Firmicutes;...
#       ... many more rows ...
#
#   The taxonomy column is not required.
#
}
