% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.getRemovedFeatures.r
\name{omicplotr.getRemovedFeatures}
\alias{omicplotr.getRemovedFeatures}
\title{Display removed features.}
\usage{
omicplotr.getRemovedFeatures(x, x.clr)
}
\arguments{
\item{x}{unfiltered data.}

\item{x.clr}{output from \code{omicplotr.clr}.}
}
\description{
Generates table of features (rows) removed by \code{omicplotr.filter}.
}
