% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.getRemovedSamples.r
\name{omicplotr.getRemovedSamples}
\alias{omicplotr.getRemovedSamples}
\title{Display removed samples.}
\usage{
omicplotr.getRemovedSamples(x, x.clr)
}
\arguments{
\item{x}{unfiltered data.}

\item{x.clr}{output from \code{omicplotr.clr}.}
}
\description{
Generates table of removed samples (columns) by \code{omicplotr.filter}.
}
