% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.permanova.r
\name{omicplotr.permanova}
\alias{omicplotr.permanova}
\title{Calculate PERMANOVA between conditions.}
\usage{
omicplotr.permanova(x, conds)
}
\arguments{
\item{x}{filtered data output from \code{omicplotr.filter}. columns must
in ordered of condition being tested (i.e., first 4 columns condition x and
next 4 columns condition 4).}

\item{conds}{a character vector of group labels.}
}
\description{
Wrapper for \code{vegan} package to calculate PERMANOVA between
 conditions from cenetered-log ratio values.
}
\examples{
for dataset where first 7 columns are condition x, next 7 are condition y.
conds <- c(rep("group1", 7). rep("group2", 7))
}
