% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.report.r
\name{omicplotr.report}
\alias{omicplotr.report}
\title{Generate summary report of significantly different features.}
\usage{
omicplotr.report(x.all, feature = NULL)
}
\arguments{
\item{x.all}{output from \code{ALDEx2}}

\item{feature}{optional; report for single feature/OTU}
}
\description{
Displays table of features found to be significantly different by
 \code{ALDEx2}, when effect size > 1 and expected Benjamini Hochberg
 corrected p-values. In progress and will show more.
}
\examples{

}
