% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.run.r
\name{omicplotr.run}
\alias{omicplotr.run}
\title{Run \code{omicplotR} \code{Shiny} app.}
\usage{
omicplotr.run()
}
\arguments{
\item{x}{Filtered data output from \code{omicplotr.filter}. Columns must
in ordered of condition being tested (i.e., first 4 columns condition x and
next 4 columns condition 4).}

\item{conds}{a character vector of group labels.}
}
\description{
Launches \code{Shiny} for \code{omicplotR} in default browser.
}
\examples{
For dataset where first 7 columns are condition x, next 7 are condition:
conds <- c(rep("group1", 7). rep("group2", 7))
}
